register validate account

+++ Controller/ConnectionController.php
@@ -163,8 +163,9 @@ class ConnectionController extends AbstractController
             $code = [rand(0, 9), rand(0, 9), rand(0, 9), rand(0, 9), rand(0, 9), rand(0, 9), rand(0, 9), rand(0, 9)];
             $code = join('', $code);
 
+            NumberManager::addNumber($user->getId(), $code);
+
             if (self::createAccountMail($user->getId())) {
-                NumberManager::addNumber($user->getId(), $code);
                 self::codePage($user->getId());
             } else {
                 $_SESSION['error'] = ["Une erreur s'est produite, veuillez réessayer ultérieurement"];
@@ -254,7 +255,13 @@ class ConnectionController extends AbstractController
             exit();
         }
 
-        if ((int) $_POST['code'] === NumberManager::getNumberByUserId($id)->getNumber() ) {
+        if (NumberManager::getNumberByUserId($id) === null) {
+            $_SESSION['error'] = ['Ce compte a déjà été vérifié'];
+            self::default();
+            exit();
+        }
+
+        if ((int) $_POST['code'] === NumberManager::getNumberByUserId($id)->getNumber()) {
             NumberManager::deleteNumber($id);
             $_SESSION['error'] = ['Compte validé, vous pouvez désormais vous connecter'];
             $_SESSION['color'] = Config::SUCCESS;
